# pkg/api/v1/gen/model.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Remote",
    "version": "v1"
  },
  "tags": [
    {
      "name": "Remote"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/service/v1/click-mouse/{code}": {
      "post": {
        "operationId": "Remote_MouseClick",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1MouseClickResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "Remote"
        ]
      }
    },
    "/service/v1/exec-task/{id}": {
      "post": {
        "operationId": "Remote_ExecTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ExecTaskResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RemoteExecTaskBody"
            }
          }
        ],
        "tags": [
          "Remote"
        ]
      }
    },
    "/service/v1/move-mouse": {
      "post": {
        "operationId": "Remote_MouseMove",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1MouseMoveResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1MouseMoveRequest"
            }
          }
        ],
        "tags": [
          "Remote"
        ]
      }
    },
    "/service/v1/position-mouse": {
      "post": {
        "operationId": "Remote_MousePosition",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1MousePositionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1MousePositionRequest"
            }
          }
        ],
        "tags": [
          "Remote"
        ]
      }
    },
    "/service/v1/press-key/{code}": {
      "post": {
        "operationId": "Remote_PressKey",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1KeyResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "Remote"
        ]
      }
    }
  },
  "definitions": {
    "RemoteExecTaskBody": {
      "type": "object",
      "properties": {
        "args": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1ExecTaskResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "results": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "v1KeyResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "v1ListTaskResponse": {
      "type": "object",
      "properties": {
        "tasks": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Task"
          }
        }
      }
    },
    "v1MouseClickResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "v1MouseMoveRequest": {
      "type": "object",
      "properties": {
        "delta": {
          "type": "integer",
          "format": "int64"
        },
        "direction": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "v1MouseMoveResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "v1MousePositionRequest": {
      "type": "object",
      "properties": {
        "x": {
          "type": "integer",
          "format": "int64"
        },
        "y": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "v1MousePositionResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "v1Task": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    }
  }
}